# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مولدتي - Iraqi Generator Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (owner only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
...
  /subscribers:
    get:
      operationId: listSubscribers
      tags: [subscribers]
      summary: List subscribers with optional search and filter
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or breaker owner name
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/SubscriberStatus"
      responses:
        "200":
          description: List of subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscriber"
    post:
      operationId: createSubscriber
      tags: [subscribers]
      summary: Create a new subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriberBody"
      responses:
        "201":
          description: Subscriber created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
  /subscribers/{id}:
    get:
      operationId: getSubscriber
      tags: [subscribers]
      summary: Get subscriber by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscriber found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
    put:
      operationId: updateSubscriber
      tags: [subscribers]
      summary: Update subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriberBody"
      responses:
        "200":
          description: Subscriber updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
    delete:
      operationId: deleteSubscriber
      tags: [subscribers]
      summary: Delete subscriber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subscriber deleted
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices with filters
      parameters:
        - name: generatorId
          in: query
          schema:
            type: integer
        - name: subscriberId
          in: query
          schema:
            type: integer
        - name: workerId
          in: query
          schema:
            type: integer
        - name: monthYear
          in: query
          schema:
            type: string
          description: "Month filter e.g. 2025-01"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/InvoiceStatus"
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create/record a payment invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/bulk-generate:
    post:
      operationId: bulkGenerateInvoices
      tags: [invoices]
      summary: Bulk generate invoices for all active subscribers for a given month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkGenerateBody"
      responses:
        "200":
          description: Bulk generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkGenerateResult"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice (record payment)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [owner, admin, worker]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        assignedGeneratorId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - username
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        assignedGeneratorId:
          type: integer
          nullable: true
      required:
        - name
        - username
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        assignedGeneratorId:
          type: integer
          nullable: true
...
    SubscriberStatus:
      type: string
      enum: [active, paused]
    Subscriber:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        breakerOwnerName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        generatorId:
          type: integer
        defaultAmperes:
          type: integer
        status:
          $ref: "#/components/schemas/SubscriberStatus"
        currentDebt:
          type: integer
          description: Current outstanding debt in IQD
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - breakerOwnerName
        - generatorId
        - defaultAmperes
        - status
        - currentDebt
        - createdAt
    CreateSubscriberBody:
      type: object
      properties:
        name:
          type: string
        breakerOwnerName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        generatorId:
          type: integer
        defaultAmperes:
          type: integer
        status:
          $ref: "#/components/schemas/SubscriberStatus"
      required:
        - name
        - breakerOwnerName
        - generatorId
        - defaultAmperes
    UpdateSubscriberBody:
      type: object
      properties:
        name:
          type: string
        breakerOwnerName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        defaultAmperes:
          type: integer
        status:
          $ref: "#/components/schemas/SubscriberStatus"
    InvoiceStatus:
      type: string
      enum: [paid, partial, unpaid]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        generatorId:
          type: integer
        workerId:
          type: integer
          nullable: true
        amperes:
          type: integer
        previousDebt:
          type: integer
        amountExpected:
          type: integer
          description: "(Amperes * Price) + Previous Debt in IQD"
        amountReceived:
          type: integer
        monthYear:
          type: string
          description: "YYYY-MM format"
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        subscriber:
          $ref: "#/components/schemas/Subscriber"
          nullable: true
        worker:
          $ref: "#/components/schemas/User"
          nullable: true
      required:
        - id
        - subscriberId
        - generatorId
        - amperes
        - previousDebt
        - amountExpected
        - amountReceived
        - monthYear
        - status
        - createdAt
    CreateInvoiceBody:
      type: object
      properties:
        subscriberId:
          type: integer
        generatorId:
          type: integer
        amperes:
          type: integer
        previousDebt:
          type: integer
          default: 0
        amountReceived:
          type: integer
        monthYear:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - subscriberId
        - generatorId
        - amperes
        - amountReceived
        - monthYear
    UpdateInvoiceBody:
      type: object
      properties:
        amperes:
          type: integer
        amountReceived:
          type: integer
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/InvoiceStatus"
    BulkGenerateBody:
      type: object
      properties:
        generatorId:
          type: integer
        monthYear:
          type: string
          description: "YYYY-MM format"
      required:
        - generatorId
        - monthYear
    BulkGenerateResult:
      type: object
      properties:
        generated:
          type: integer
        skipped:
          type: integer
        monthYear:
          type: string
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
      required:
        - generated
        - skipped
        - monthYear
        - invoices
...